refactor: M3-6195 - Simplify import path
