+++ packages/manager/src/features/ObjectStorage/BucketDetail/CreateFolderDrawer.tsx
@@ -4,7 +4,7 @@ import TextField from 'src/components/TextField';
 import ActionsPanel from 'src/components/ActionsPanel';
 import { useFormik } from 'formik';
 import { useCreateObjectUrlMutation } from 'src/queries/objectStorage';
-import Button from 'src/components/Button/Button';
+import Button from 'src/components/Button';
 
 interface Props {
   open: boolean;
